Moved app to bottom navigation
