+++ app/build.gradle
@@ -65,4 +65,8 @@ dependencies {
 
     implementation 'com.github.andreilisun:swipedismissdialog:0.1'
     implementation 'com.android.support:cardview-v7:28.0.0'
+
+    implementation 'com.github.stfalcon:stfalcon-imageviewer:1.0.0'
+    implementation 'androidx.navigation:navigation-fragment:2.0.0'
+    implementation 'androidx.navigation:navigation-ui:2.0.0'
 }
