+++ app/src/main/java/in/ac/bits_hyderabad/swd/swd/user/activity/OrderGoodie.java
@@ -7,19 +7,9 @@ import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.Uri;
-
-import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import androidx.appcompat.widget.Toolbar;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.provider.ContactsContract;
-import android.util.DisplayMetrics;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
-import android.view.Window;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.CheckBox;
@@ -31,6 +21,9 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
@@ -38,23 +31,17 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.github.andreilisun.swipedismissdialog.SwipeDismissDialog;
-import com.google.gson.JsonObject;
-import com.itextpdf.text.pdf.parser.Line;
 import com.jsibbold.zoomage.ZoomageView;
 import com.squareup.picasso.Picasso;
+import com.stfalcon.imageviewer.StfalconImageViewer;
+import com.stfalcon.imageviewer.loader.ImageLoader;
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.lang.reflect.Array;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Iterator;
-import java.util.Locale;
 import java.util.Map;
 
 import in.ac.bits_hyderabad.swd.swd.R;
@@ -159,7 +146,8 @@ public class OrderGoodie extends AppCompatActivity {
         Picasso.get().load("http://swd.bits-hyderabad.ac.in/goodies/img/" + goodie.getImage())
                 .resize(125, 125)
                 .placeholder(R.drawable.ic_loading)
-                .centerInside().error(R.drawable.ic_error)
+                .centerInside()
+                .error(R.drawable.ic_error)
                 .into(ivImageOrder);
 
 
@@ -192,27 +180,16 @@ public class OrderGoodie extends AppCompatActivity {
         ivImageOrder.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                ImageView image;
-                Dialog myDialog = new Dialog(OrderGoodie.this, android.R.style.Theme_Translucent_NoTitleBar_Fullscreen);
-                myDialog.setContentView(R.layout.dialog_my);
-                myDialog.getWindow().setBackgroundDrawableResource(R.color.semiTransparentColor99black);
-                image = myDialog.findViewById(R.id.ivFullGoodieImage);
-
-
-                String ImageUrl = "http://swd.bits-hyderabad.ac.in/goodies/img/" + goodie.getImage();
-
-                Picasso.get().load(ImageUrl)
-                        .resize(1500, 1500)
-                        .centerInside()
-                        .placeholder(R.drawable.ic_loading)
-                        .error(R.drawable.ic_error)
-                        .into(image);
-                myDialog.show();
-
+                String[] ImageUrl = {"http://swd.bits-hyderabad.ac.in/goodies/img/" + goodie.getImage()};
+                new StfalconImageViewer.Builder<String>(getApplicationContext(), ImageUrl, new ImageLoader<String>() {
+                    @Override
+                    public void loadImage(ImageView imageView, String image) {
+                        Picasso.get().load(image).into(imageView);
+                    }
+                }).show();
             }
         });
 
-        //KYA HI CHUTIYAAPP HAI YE
         if (goodie.getXs().equals("0")) {
             rlxs.setVisibility(View.GONE);
         }
