+++ app/src/main/res/color/bottom_navigation_colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/green_accent" android:state_checked="true" />
+    <item android:color="@color/whiteText" android:state_checked="false" />
+</selector>
