+++ app/src/main/res/color/profile_input_layout_colors.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/green_accent" android:state_focused="true" />
+    <item android:alpha="0.87" android:color="@color/whiteText" android:state_hovered="true" />
+    <item android:alpha="0.12" android:color="@color/whiteText" android:state_enabled="false" />
+    <item android:alpha="0.38" android:color="@color/whiteText" />
+</selector>
