+++ app/src/main/res/color/profile_input_layout_text_colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/green_accent" android:state_enabled="false" />
+    <item android:alpha="0.6" android:color="@color/whiteText" />
+</selector>
