+++ app/src/main/res/layout/activity_order_goodie.xml
@@ -2,6 +2,7 @@
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:theme="@style/Theme.AppCompat"
     android:layout_height="wrap_content"
     android:layout_width="match_parent" >
 <LinearLayout
