+++ app/src/main/res/layout/fragment_home.xml
@@ -6,23 +6,6 @@
     android:layout_width="match_parent"
     android:background="@color/dark_background">
 
-    <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-        android:id="@+id/fabContactUs"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="16dp"
-        android:backgroundTint="@color/green_accent"
-        android:elevation="10dp"
-        android:paddingHorizontal="16dp"
-        android:paddingVertical="12dp"
-        android:text="@string/contact"
-        android:textColor="@color/whiteText"
-        app:icon="@drawable/outline_email_24"
-        app:iconTint="@color/whiteText" />
-
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -34,9 +17,8 @@
             android:layout_marginStart="16dp"
             android:layout_marginTop="16dp"
             android:layout_marginEnd="16dp"
-            android:elevation="0dp"
-            app:cardBackgroundColor="@color/background_overlay"
             app:cardCornerRadius="8dp"
+            app:cardBackgroundColor="@color/dark_background"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/divider">
@@ -48,8 +30,8 @@
 
                 <ImageView
                     android:id="@+id/imageView4"
-                    android:layout_width="40dp"
-                    android:layout_height="40dp"
+                    android:layout_width="32dp"
+                    android:layout_height="32dp"
                     android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:layout_marginBottom="16dp"
@@ -64,7 +46,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="16dp"
                     android:layout_marginEnd="4dp"
-                    android:fontFamily="sans-serif-black"
+                    android:fontFamily="sans-serif"
                     android:text="@string/timetable_division"
                     android:textAppearance="@style/TextAppearance.AppCompat.Display1"
                     android:textColor="@color/whiteText"
@@ -95,9 +77,9 @@
             android:layout_marginStart="16dp"
             android:layout_marginTop="16dp"
             android:layout_marginEnd="16dp"
-            android:elevation="0dp"
-            app:cardBackgroundColor="@color/background_overlay"
+            android:elevation="10dp"
             app:cardCornerRadius="8dp"
+            app:cardBackgroundColor="@color/dark_background"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/cvTD">
@@ -109,8 +91,8 @@
 
                 <ImageView
                     android:id="@+id/imageView6"
-                    android:layout_width="40dp"
-                    android:layout_height="40dp"
+                    android:layout_width="32dp"
+                    android:layout_height="32dp"
                     android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:layout_marginBottom="16dp"
@@ -126,7 +108,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="16dp"
                     android:layout_marginEnd="4dp"
-                    android:fontFamily="sans-serif-black"
+                    android:fontFamily="sans-serif"
                     android:text="@string/library_opac"
                     android:textAppearance="@style/TextAppearance.AppCompat.Display1"
                     android:textColor="@color/whiteText"
@@ -157,9 +139,9 @@
             android:layout_marginStart="16dp"
             android:layout_marginTop="16dp"
             android:layout_marginEnd="16dp"
-            android:elevation="0dp"
-            app:cardBackgroundColor="@color/background_overlay"
+            android:elevation="10dp"
             app:cardCornerRadius="8dp"
+            app:cardBackgroundColor="@color/dark_background"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/cvOPAC">
@@ -171,8 +153,8 @@
 
                 <ImageView
                     android:id="@+id/imageView8"
-                    android:layout_width="40dp"
-                    android:layout_height="40dp"
+                    android:layout_width="32dp"
+                    android:layout_height="32dp"
                     android:layout_marginStart="16dp"
                     android:layout_marginTop="16dp"
                     android:layout_marginBottom="16dp"
@@ -188,7 +170,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="16dp"
                     android:layout_marginEnd="4dp"
-                    android:fontFamily="sans-serif-black"
+                    android:fontFamily="sans-serif"
                     android:text="@string/erp"
                     android:textAppearance="@style/TextAppearance.AppCompat.Display1"
                     android:textColor="@color/whiteText"
@@ -219,8 +201,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.4" />
+            app:layout_constraintTop_toTopOf="parent" />
 
         <ImageView
             android:id="@+id/imageView2"
@@ -233,8 +214,19 @@
             app:layout_constraintBottom_toTopOf="@id/divider"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/myProfileImageView"
             app:srcCompat="@drawable/logo" />
+
+        <ImageView
+            android:id="@+id/myProfileImageView"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:layout_marginTop="16dp"
+            android:layout_marginEnd="16dp"
+            android:tint="@color/whiteText"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@drawable/outline_account_circle_24" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
