+++ app/src/main/res/layout/nav_app_bar.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".user.activity.User_Nav">
+        tools:context=".user.activity.MainActivity">
 
 
         <com.google.android.material.appbar.AppBarLayout
