+++ app/src/main/res/layout/nav_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".user.activity.User_Nav"
+    tools:context=".user.activity.MainActivity"
     tools:showIn="@layout/nav_app_bar">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
