+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_goodies"
+        android:icon="@drawable/outline_shopping_cart_24"
+        android:title="@string/goodies" />
+
+    <item
+        android:id="@+id/navigation_mess"
+        android:icon="@drawable/outline_restaurant_menu_24"
+        android:title="@string/mess" />
+
+    <item
+        android:id="@+id/navigation_home"
+        android:icon="@drawable/outline_home_24"
+        android:title="@string/home" />
+
+    <item
+        android:id="@+id/navigation_connect"
+        android:icon="@drawable/outline_message_24"
+        android:title="@string/connect" />
+
+    <item
+        android:id="@+id/navigation_miscellaneous"
+        android:icon="@drawable/outline_more_horiz_24"
+        android:title="@string/miscellaneous" />
+
+</menu>
