+++ app/src/main/res/values/strings.xml
@@ -203,6 +203,7 @@
     <string name="important_links">Important links</string>
     <string name="timetable_division">Timetable division</string>
     <string name="erp">ERP</string>
+    <string name="miscellaneous">Miscellaneous</string>
 
 
 </resources>
